#109 Moved provider interfaces into AdminAssistant
