+++ src/Framework/AdminAssistant.Framework/Providers/DateTimeProvider.cs
@@ -2,15 +2,7 @@ using System;
 
 namespace AdminAssistant.Framework.Providers
 {
-    public interface IDateTimeProvider
-    {
-        ///<summary>
-        ///Gets a System.DateTime object that is set to the current date and time on this
-        ///computer, expressed as the Coordinated Universal Time (UTC).
-        ///</summary>
-        ///<returns>An object whose value is the current UTC date and time.</returns>
-        DateTime UtcNow { get; }
-    }
+
     [System.Diagnostics.CodeAnalysis.SuppressMessage("Build", "CA1812", Justification = "Compiler dosen't understand dependency injection")]
     internal class DateTimeProvider : IDateTimeProvider
     {
